refactor(agents): rename CommitAIPath to AgentPath (temp typealias)
